backend/requirements.txt
alembic==1.16.4
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic[email]
python-dotenv==1.0.1
python-json-logger==2.0.7
PyJWT==2.10.1
PyPDF2==3.0.1
pyotp==2.9.0
reportlab==4.4.2
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
